Move chronograph tests mixins to v2 and v3 awarding plugins
- Delete payment tests for v3 awarding
- Move switch_..._to_unsuccessful_2 chronograph tests from dgf to v2 and
v3 awarding plugins
- Delete switch_payment_to_unsuccessful_2 test for v2 awarding plugin
- Add condition in switch_..._to_unsuccessful_2 test
 for procurementMethodType
- Remove chronograph.py module with mixins from core tests
