+++ openprocurement/auctions/core/plugins/awarding/v2/tests/chronograph.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 from openprocurement.auctions.core.tests.base import snitch
-from openprocurement.auctions.core.tests.blanks.chronograph_blanks import (
+from openprocurement.auctions.core.plugins.awarding.v2.tests.blanks.chronograph_blanks import (
+    # AuctionAwardSwitchResourceTest
     switch_verification_to_unsuccessful,
     switch_payment_to_unsuccessful,
     switch_active_to_unsuccessful,
+    # AuctionDontSwitchSuspendedAuctionResourceTest
     switch_suspended_verification_to_unsuccessful,
     switch_suspended_payment_to_unsuccessful,
     switch_suspended_active_to_unsuccessful
