+++ openprocurement/auctions/core/plugins/awarding/v3/tests/blanks/chronograph_blanks.py
@@ -1,0 +1,207 @@
+# -*- coding: utf-8 -*-
+# AuctionAwardSwitchResourceTest
+
+
+def switch_verification_to_unsuccessful(self):
+    auction = self.db.get(self.auction_id)
+    auction['awards'][0]['verificationPeriod']['endDate'] = auction['awards'][0]['verificationPeriod']['startDate']
+    self.db.save(auction)
+
+    self.app.authorization = ('Basic', ('chronograph', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
+    self.assertEqual(response.status, '200 OK')
+    auction = response.json['data']
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
+    self.assertEqual(auction['awards'][1]['status'], 'pending')
+    self.assertEqual(auction['status'], 'active.qualification')
+    self.assertNotIn('endDate', auction['awardPeriod'])
+
+
+def switch_active_to_unsuccessful(self):
+    bid_token = self.initial_bids_tokens[self.award['bid_id']]
+    response = self.app.post('/auctions/{}/awards/{}/documents?acc_token={}'.format(
+        self.auction_id, self.award_id, self.auction_token), upload_files=[('file', 'auction_protocol.pdf', 'content')])
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    doc_id = response.json["data"]['id']
+    key = response.json["data"]["url"].split('?')[-1]
+
+    response = self.app.patch_json(
+        '/auctions/{}/awards/{}/documents/{}?acc_token={}'.format(self.auction_id, self.award_id, doc_id,
+                                                                  self.auction_token), {"data": {
+            "description": "auction protocol",
+            "documentType": 'auctionProtocol'
+        }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json["data"]["documentType"], 'auctionProtocol')
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, self.award_id),
+                                   {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], "active")
+
+    auction = self.db.get(self.auction_id)
+    auction['awards'][0]['signingPeriod']['endDate'] = auction['awards'][0]['signingPeriod']['startDate']
+    self.db.save(auction)
+
+    self.app.authorization = ('Basic', ('chronograph', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
+    auction = response.json['data']
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
+    self.assertEqual(auction['contracts'][0]['status'], 'cancelled')
+    self.assertEqual(auction['awards'][1]['status'], 'pending')
+    self.assertEqual(auction['status'], 'active.qualification')
+    self.assertNotIn('endDate', auction['awardPeriod'])
+
+# AuctionDontSwitchSuspendedAuctionResourceTest
+
+
+def switch_suspended_verification_to_unsuccessful(self):
+    auction = self.db.get(self.auction_id)
+    auction['awards'][0]['verificationPeriod']['endDate'] = auction['awards'][0]['verificationPeriod']['startDate']
+    self.db.save(auction)
+
+    self.app.authorization = ('Basic', ('administrator', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': True}})
+
+    self.app.authorization = ('Basic', ('chronograph', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
+    self.assertEqual(response.status, '200 OK')
+    auction = response.json['data']
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(auction['awards'][0]['status'], 'pending')
+    self.assertEqual(auction['awards'][1]['status'], 'pending.waiting')
+
+    self.app.authorization = ('Basic', ('administrator', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': False}})
+
+    self.app.authorization = ('Basic', ('chronograph', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
+    self.assertEqual(response.status, '200 OK')
+    auction = response.json['data']
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
+    self.assertEqual(auction['awards'][1]['status'], 'pending')
+    self.assertEqual(auction['status'], 'active.qualification')
+    self.assertNotIn('endDate', auction['awardPeriod'])
+
+
+def switch_suspended_active_to_unsuccessful(self):
+    bid_token = self.initial_bids_tokens[self.award['bid_id']]
+    response = self.app.post('/auctions/{}/awards/{}/documents?acc_token={}'.format(
+        self.auction_id, self.award_id, self.auction_token), upload_files=[('file', 'auction_protocol.pdf', 'content')])
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    doc_id = response.json["data"]['id']
+    key = response.json["data"]["url"].split('?')[-1]
+
+    response = self.app.patch_json(
+        '/auctions/{}/awards/{}/documents/{}?acc_token={}'.format(self.auction_id, self.award_id, doc_id,
+                                                                  self.auction_token), {"data": {
+            "description": "auction protocol",
+            "documentType": 'auctionProtocol'
+        }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json["data"]["documentType"], 'auctionProtocol')
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, self.award_id),
+                                   {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], "active")
+
+    auction = self.db.get(self.auction_id)
+    auction['awards'][0]['signingPeriod']['endDate'] = auction['awards'][0]['signingPeriod']['startDate']
+    self.db.save(auction)
+
+    self.app.authorization = ('Basic', ('administrator', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': True}})
+
+    self.app.authorization = ('Basic', ('chronograph', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
+    self.assertEqual(response.status, '200 OK')
+    auction = response.json['data']
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(auction['awards'][0]['status'], 'active')
+    self.assertEqual(auction['contracts'][0]['status'], 'pending')
+    self.assertEqual(auction['awards'][1]['status'], 'pending.waiting')
+    self.assertEqual(auction['status'], 'active.awarded')
+    self.assertIn('endDate', auction['awardPeriod'])
+
+    self.app.authorization = ('Basic', ('administrator', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': False}})
+
+    self.app.authorization = ('Basic', ('chronograph', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
+    auction = response.json['data']
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
+    self.assertEqual(auction['contracts'][0]['status'], 'cancelled')
+    self.assertEqual(auction['awards'][1]['status'], 'pending')
+    self.assertEqual(auction['status'], 'active.qualification')
+    self.assertNotIn('endDate', auction['awardPeriod'])
+
+# AuctionAwardSwitch2ResourceTest
+
+
+def switch_verification_to_unsuccessful_2(self):
+    auction = self.db.get(self.auction_id)
+    auction['awards'][0]['verificationPeriod']['endDate'] = auction['awards'][0]['verificationPeriod']['startDate']
+    self.db.save(auction)
+
+    self.app.authorization = ('Basic', ('chronograph', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
+    self.assertEqual(response.status, '200 OK')
+    auction = response.json['data']
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
+    if 'Insider' not in auction['procurementMethodType']:
+        self.assertEqual(auction['awards'][1]['status'], 'unsuccessful')
+        self.assertEqual(auction['status'], 'unsuccessful')
+        self.assertIn('endDate', auction['awardPeriod'])
+
+
+def switch_active_to_unsuccessful_2(self):
+    bid_token = self.initial_bids_tokens[self.award['bid_id']]
+    response = self.app.post('/auctions/{}/awards/{}/documents?acc_token={}'.format(
+        self.auction_id, self.award_id, self.auction_token), upload_files=[('file', 'auction_protocol.pdf', 'content')])
+    self.assertEqual(response.status, '201 Created')
+    self.assertEqual(response.content_type, 'application/json')
+    doc_id = response.json["data"]['id']
+    key = response.json["data"]["url"].split('?')[-1]
+
+    response = self.app.patch_json(
+        '/auctions/{}/awards/{}/documents/{}?acc_token={}'.format(self.auction_id, self.award_id, doc_id,
+                                                                  self.auction_token), {"data": {
+            "description": "auction protocol",
+            "documentType": 'auctionProtocol'
+        }})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json["data"]["documentType"], 'auctionProtocol')
+
+    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, self.award_id),
+                                   {"data": {"status": "active"}})
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(response.content_type, 'application/json')
+    self.assertEqual(response.json['data']["status"], "active")
+
+    auction = self.db.get(self.auction_id)
+    auction['awards'][0]['signingPeriod']['endDate'] = auction['awards'][0]['signingPeriod']['startDate']
+    self.db.save(auction)
+
+    self.app.authorization = ('Basic', ('chronograph', ''))
+    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
+    auction = response.json['data']
+    self.assertEqual(response.status, '200 OK')
+    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
+    self.assertEqual(auction['contracts'][0]['status'], 'cancelled')
+    if 'Insider' not in auction['procurementMethodType']:
+        self.assertEqual(auction['awards'][1]['status'], 'unsuccessful')
+        self.assertEqual(auction['status'], 'unsuccessful')
+        self.assertIn('endDate', auction['awardPeriod'])
