+++ openprocurement/auctions/core/tests/blanks/chronograph_blanks.py
@@ -28,94 +28,6 @@ def switch_to_unsuccessful(self):
     if self.initial_lots:
         self.assertEqual(set([i['status'] for i in response.json['data']["lots"]]), set(["unsuccessful"]))
 
-# AuctionAwardSwitchResourceTest
-
-
-def switch_verification_to_unsuccessful(self):
-    auction = self.db.get(self.auction_id)
-    auction['awards'][0]['verificationPeriod']['endDate'] = auction['awards'][0]['verificationPeriod']['startDate']
-    self.db.save(auction)
-
-    self.app.authorization = ('Basic', ('chronograph', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
-    self.assertEqual(response.status, '200 OK')
-    auction = response.json['data']
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
-    self.assertEqual(auction['awards'][1]['status'], 'pending')
-    self.assertEqual(auction['status'], 'active.qualification')
-    self.assertNotIn('endDate', auction['awardPeriod'])
-
-
-def switch_payment_to_unsuccessful(self):
-    bid_token = self.initial_bids_tokens[self.award['bid_id']]
-    response = self.app.post('/auctions/{}/awards/{}/documents?acc_token={}'.format(
-        self.auction_id, self.award_id, self.auction_token), upload_files=[('file', 'auction_protocol.pdf', 'content')])
-    self.assertEqual(response.status, '201 Created')
-    self.assertEqual(response.content_type, 'application/json')
-    doc_id = response.json["data"]['id']
-    key = response.json["data"]["url"].split('?')[-1]
-
-    response = self.app.patch_json(
-        '/auctions/{}/awards/{}/documents/{}?acc_token={}'.format(self.auction_id, self.award_id, doc_id,
-                                                                  self.auction_token), {"data": {
-            "description": "auction protocol",
-            "documentType": 'auctionProtocol'
-        }})
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(response.content_type, 'application/json')
-    self.assertEqual(response.json["data"]["documentType"], 'auctionProtocol')
-
-    self.app.authorization = ('Basic', ('chronograph', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
-    self.assertEqual(response.status, '200 OK')
-    auction = response.json['data']
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(auction['awards'][0]['status'], 'pending')
-    self.assertEqual(auction['awards'][1]['status'], 'pending.waiting')
-    self.assertEqual(auction['status'], 'active.qualification')
-    self.assertNotIn('endDate', auction['awardPeriod'])
-
-
-def switch_active_to_unsuccessful(self):
-    bid_token = self.initial_bids_tokens[self.award['bid_id']]
-    response = self.app.post('/auctions/{}/awards/{}/documents?acc_token={}'.format(
-        self.auction_id, self.award_id, self.auction_token), upload_files=[('file', 'auction_protocol.pdf', 'content')])
-    self.assertEqual(response.status, '201 Created')
-    self.assertEqual(response.content_type, 'application/json')
-    doc_id = response.json["data"]['id']
-    key = response.json["data"]["url"].split('?')[-1]
-
-    response = self.app.patch_json(
-        '/auctions/{}/awards/{}/documents/{}?acc_token={}'.format(self.auction_id, self.award_id, doc_id,
-                                                                  self.auction_token), {"data": {
-            "description": "auction protocol",
-            "documentType": 'auctionProtocol'
-        }})
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(response.content_type, 'application/json')
-    self.assertEqual(response.json["data"]["documentType"], 'auctionProtocol')
-
-    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, self.award_id),
-                                   {"data": {"status": "active"}})
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(response.content_type, 'application/json')
-    self.assertEqual(response.json['data']["status"], "active")
-
-    auction = self.db.get(self.auction_id)
-    auction['awards'][0]['signingPeriod']['endDate'] = auction['awards'][0]['signingPeriod']['startDate']
-    self.db.save(auction)
-
-    self.app.authorization = ('Basic', ('chronograph', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
-    auction = response.json['data']
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
-    self.assertEqual(auction['contracts'][0]['status'], 'cancelled')
-    self.assertEqual(auction['awards'][1]['status'], 'pending')
-    self.assertEqual(auction['status'], 'active.qualification')
-    self.assertNotIn('endDate', auction['awardPeriod'])
-
 # AuctionComplaintSwitchResourceTest
 
 
@@ -269,137 +181,3 @@ def switch_suspended_auction_to_auction(self):
     self.assertEqual(response.status, '200 OK')
     self.assertEqual(response.content_type, 'application/json')
     self.assertEqual(response.json['data']["status"], "active.auction")
-
-# AuctionDontSwitchSuspendedAuctionResourceTest
-
-
-def switch_suspended_verification_to_unsuccessful(self):
-    auction = self.db.get(self.auction_id)
-    auction['awards'][0]['verificationPeriod']['endDate'] = auction['awards'][0]['verificationPeriod']['startDate']
-    self.db.save(auction)
-
-    self.app.authorization = ('Basic', ('administrator', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': True}})
-
-    self.app.authorization = ('Basic', ('chronograph', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
-    self.assertEqual(response.status, '200 OK')
-    auction = response.json['data']
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(auction['awards'][0]['status'], 'pending')
-    self.assertEqual(auction['awards'][1]['status'], 'pending.waiting')
-
-    self.app.authorization = ('Basic', ('administrator', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': False}})
-
-    self.app.authorization = ('Basic', ('chronograph', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
-    self.assertEqual(response.status, '200 OK')
-    auction = response.json['data']
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
-    self.assertEqual(auction['awards'][1]['status'], 'pending')
-    self.assertEqual(auction['status'], 'active.qualification')
-    self.assertNotIn('endDate', auction['awardPeriod'])
-
-
-def switch_suspended_payment_to_unsuccessful(self):
-    bid_token = self.initial_bids_tokens[self.award['bid_id']]
-    response = self.app.post('/auctions/{}/awards/{}/documents?acc_token={}'.format(
-        self.auction_id, self.award_id, self.auction_token), upload_files=[('file', 'auction_protocol.pdf', 'content')])
-    self.assertEqual(response.status, '201 Created')
-    self.assertEqual(response.content_type, 'application/json')
-    doc_id = response.json["data"]['id']
-    key = response.json["data"]["url"].split('?')[-1]
-
-    response = self.app.patch_json(
-        '/auctions/{}/awards/{}/documents/{}?acc_token={}'.format(self.auction_id, self.award_id, doc_id,
-                                                                  self.auction_token), {"data": {
-            "description": "auction protocol",
-            "documentType": 'auctionProtocol'
-        }})
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(response.content_type, 'application/json')
-    self.assertEqual(response.json["data"]["documentType"], 'auctionProtocol')
-
-    # XXX TODO FIX TEST
-    # self.app.authorization = ('Basic', ('administrator', ''))
-    # response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': True}})
-    #
-    # self.app.authorization = ('Basic', ('chronograph', ''))
-    # response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
-    # self.assertEqual(response.status, '200 OK')
-    # auction = response.json['data']
-    # self.assertEqual(response.status, '200 OK')
-    # self.assertEqual(auction['awards'][0]['status'], 'pending')
-    # self.assertEqual(auction['awards'][1]['status'], 'pending.waiting')
-    #
-    # self.app.authorization = ('Basic', ('administrator', ''))
-    # response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': False}})
-    #
-    # self.app.authorization = ('Basic', ('chronograph', ''))
-    # response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
-    # self.assertEqual(response.status, '200 OK')
-    # auction = response.json['data']
-    # self.assertEqual(response.status, '200 OK')
-    # self.assertEqual(auction['awards'][0]['status'], 'pending')
-    # self.assertEqual(auction['awards'][1]['status'], 'pending.waiting')
-    # self.assertEqual(auction['status'], 'active.qualification')
-    # self.assertNotIn('endDate', auction['awardPeriod'])
-
-
-def switch_suspended_active_to_unsuccessful(self):
-    bid_token = self.initial_bids_tokens[self.award['bid_id']]
-    response = self.app.post('/auctions/{}/awards/{}/documents?acc_token={}'.format(
-        self.auction_id, self.award_id, self.auction_token), upload_files=[('file', 'auction_protocol.pdf', 'content')])
-    self.assertEqual(response.status, '201 Created')
-    self.assertEqual(response.content_type, 'application/json')
-    doc_id = response.json["data"]['id']
-    key = response.json["data"]["url"].split('?')[-1]
-
-    response = self.app.patch_json(
-        '/auctions/{}/awards/{}/documents/{}?acc_token={}'.format(self.auction_id, self.award_id, doc_id,
-                                                                  self.auction_token), {"data": {
-            "description": "auction protocol",
-            "documentType": 'auctionProtocol'
-        }})
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(response.content_type, 'application/json')
-    self.assertEqual(response.json["data"]["documentType"], 'auctionProtocol')
-
-    response = self.app.patch_json('/auctions/{}/awards/{}'.format(self.auction_id, self.award_id),
-                                   {"data": {"status": "active"}})
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(response.content_type, 'application/json')
-    self.assertEqual(response.json['data']["status"], "active")
-
-    auction = self.db.get(self.auction_id)
-    auction['awards'][0]['signingPeriod']['endDate'] = auction['awards'][0]['signingPeriod']['startDate']
-    self.db.save(auction)
-
-    self.app.authorization = ('Basic', ('administrator', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': True}})
-
-    self.app.authorization = ('Basic', ('chronograph', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
-    self.assertEqual(response.status, '200 OK')
-    auction = response.json['data']
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(auction['awards'][0]['status'], 'active')
-    self.assertEqual(auction['contracts'][0]['status'], 'pending')
-    self.assertEqual(auction['awards'][1]['status'], 'pending.waiting')
-    self.assertEqual(auction['status'], 'active.awarded')
-    self.assertIn('endDate', auction['awardPeriod'])
-
-    self.app.authorization = ('Basic', ('administrator', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'suspended': False}})
-
-    self.app.authorization = ('Basic', ('chronograph', ''))
-    response = self.app.patch_json('/auctions/{}'.format(self.auction_id), {'data': {'id': self.auction_id}})
-    auction = response.json['data']
-    self.assertEqual(response.status, '200 OK')
-    self.assertEqual(auction['awards'][0]['status'], 'unsuccessful')
-    self.assertEqual(auction['contracts'][0]['status'], 'cancelled')
-    self.assertEqual(auction['awards'][1]['status'], 'pending')
-    self.assertEqual(auction['status'], 'active.qualification')
-    self.assertNotIn('endDate', auction['awardPeriod'])
